task2/requirements.txt
torch
torchvision
transformers
datasets
accelerate
scikit-learn
numpy
matplotlib
seaborn
pandas
Pillow
tqdm
jupyter
kaggle
